backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiohttp>=3.10.0
aiofiles>=23.2.1
pillow>=10.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
